EditAccountRequests and some major edits for bugs
first Bugs: make account had an error with the date because of the datatype in the database so it was always filled with null \n the unique identifier of the account was 8 numbers which caused validation error so it was changed to 5 numbers to match the business logic \n ValidateAccounts: the comments were not imported in the field so any comment was just overwriting the comment of the record instead of just editing it now checker can see the comments of the maker and edit upon it if he/she required another edit \n bining the data in ValidateAccount was failing verfication for partial updated using the Attach() method so this changed into fetching the whole row , editing it and then updating the database with the new row \n finally the new feature : Account Edit requst was made such that fields with lookup tables (currency, class,branch) will be in ddl the currency is implemented and the other two filds will be implemented in the next commit and also lookup tables were cached boosting the performance dueto hitting the data less with a static/application variable so that it is imported once and used alot

+++ currency_rate_look_up.cs
@@ -14,9 +14,9 @@ namespace NBE
     
     public partial class currency_rate_look_up
     {
-        public float RateID { get; set; }
+        public int RateID { get; set; }
         public string FromCur { get; set; }
         public string ToCur { get; set; }
-        public int Rate { get; set; }
+        public double Rate { get; set; }
     }
 }
